+++ src/main/java/balancetalk/member/dto/ApiMember.java
@@ -1,0 +1,20 @@
+package balancetalk.member.dto;
+
+import balancetalk.global.exception.BalanceTalkException;
+import balancetalk.global.exception.ErrorCode;
+import balancetalk.member.domain.Member;
+import balancetalk.member.domain.MemberRepository;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+
+@Data
+@AllArgsConstructor
+public class ApiMember {
+
+    private String username;
+
+    public Member toMember(MemberRepository memberRepository) {
+        return memberRepository.findByEmail(username)
+                .orElseThrow(() -> new BalanceTalkException(ErrorCode.NOT_FOUND_MEMBER));
+    }
+}
